Fix SDL2 compile-errors, made it work on Win64
.. when compiled with MSVC2012

+++ neo/sys/win32/win_main.cpp
@@ -393,7 +393,7 @@ Sys_ListFiles
 int Sys_ListFiles( const char *directory, const char *extension, idStrList &list ) {
 	idStr		search;
 	struct _finddata_t findinfo;
-	int			findhandle;
+	intptr_t	findhandle;
 	int			flag;
 
 	if ( !extension) {
